+++ internal/scale/scalecommon/interfaces.go
@@ -36,6 +36,10 @@ type Configuration interface {
 		container *v1.Container,
 	) error
 
+	ValidateRequestsLimits(
+		container *v1.Container,
+	) error
+
 	String() string
 }
 
@@ -53,7 +57,9 @@ type Configurations interface {
 		container *v1.Container,
 	) error
 
-	ValidateCollection() error
+	ValidateCollection(
+		container *v1.Container,
+	) error
 
 	ConfigurationFor(
 		resourceName v1.ResourceName,