+++ test/integration/integration_test.go
@@ -571,12 +571,12 @@ func TestDeploymentScaleWhenUnknownResourcesAll(t *testing.T) {
 	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
 
 	annotations := csaQuantityAnnotations{
-		cpuStartup:                "200m",
-		cpuPostStartupRequests:    "50m",
-		cpuPostStartupLimits:      "50m",
-		memoryStartup:             "200M",
-		memoryPostStartupRequests: "150M",
-		memoryPostStartupLimits:   "150M",
+		cpuStartup:                "150m",
+		cpuPostStartupRequests:    "100m",
+		cpuPostStartupLimits:      "100m",
+		memoryStartup:             "150M",
+		memoryPostStartupRequests: "100M",
+		memoryPostStartupLimits:   "100M",
 	}
 
 	config := echoDeploymentConfigStandard(
@@ -628,9 +628,9 @@ func TestDeploymentScaleWhenUnknownResourcesCpu(t *testing.T) {
 	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
 
 	annotations := csaQuantityAnnotations{
-		cpuStartup:             "200m",
-		cpuPostStartupRequests: "50m",
-		cpuPostStartupLimits:   "50m",
+		cpuStartup:             "150m",
+		cpuPostStartupRequests: "100m",
+		cpuPostStartupLimits:   "100m",
 	}
 
 	config := echoDeploymentConfigStandard(
@@ -682,9 +682,9 @@ func TestDeploymentScaleWhenUnknownResourcesMemory(t *testing.T) {
 	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
 
 	annotations := csaQuantityAnnotations{
-		memoryStartup:             "200M",
-		memoryPostStartupRequests: "150M",
-		memoryPostStartupLimits:   "150M",
+		memoryStartup:             "150M",
+		memoryPostStartupRequests: "100M",
+		memoryPostStartupLimits:   "100M",
 	}
 
 	config := echoDeploymentConfigStandard(
@@ -864,12 +864,12 @@ func TestValidationFailure(t *testing.T) {
 	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
 
 	annotations := csaQuantityAnnotations{
-		cpuStartup:                "50m",
-		cpuPostStartupRequests:    "200m",
-		cpuPostStartupLimits:      "200m",
-		memoryStartup:             "200M",
-		memoryPostStartupRequests: "150M",
-		memoryPostStartupLimits:   "150M",
+		cpuStartup:                "100m",
+		cpuPostStartupRequests:    "150m",
+		cpuPostStartupLimits:      "150m",
+		memoryStartup:             "150M",
+		memoryPostStartupRequests: "100M",
+		memoryPostStartupLimits:   "100M",
 	}
 
 	config := echoDeploymentConfigStandardStartup(namespace, 2, annotations)
@@ -884,7 +884,7 @@ func TestValidationFailure(t *testing.T) {
 	}
 
 	for _, statusAnn := range podStatusAnn {
-		assert.Contains(t, statusAnn.Status, "cpu post-startup requests (200m) is greater than startup value (50m)")
+		assert.Contains(t, statusAnn.Status, "cpu post-startup requests (150m) is greater than startup value (100m)")
 		require.NotEmpty(t, statusAnn.LastUpdated)
 
 		require.Equal(t, podcommon.StateBoolUnknown, statusAnn.States.StartupProbe)